Use cardDetailView
