+++ scripts/index.js
@@ -6,7 +6,7 @@ var _ = require('underscore'),
   AppView = require('./app_view'),
   UserView = require('./user_view'),
   FolderDetailView = require('./folder_detail_view'),
-  FolderPopView = require('./folder_pop_view'),
+  CardDetailView = require('./card_detail_view'),
   HomeView = require('./homepage/home_view'),
   Route = ReactRouter.Route,
   DefaultRoute = ReactRouter.DefaultRoute,
@@ -19,7 +19,7 @@ $(document).ready(function(){
       <Route name="user" path="/:userId" handler={UserView}>
         <Route name="folderDetail" path="/:userId/folder/:folderName" handler={FolderDetailView}/>
       </Route>
-      <Route name="folderPop" path="/:userId/folder/:folderName/card/:cardId" handler={FolderPopView}/>
+      <Route name="cardDetail" path="/:userId/folder/:folderName/card/:cardId" handler={CardDetailView}/>
     </Route>
   );
